+++ CHANGELOG.md
@@ -7,6 +7,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 Unreleased is the current development version.
 
+- Teleconnections diagnostic improvements (#672)
+
 ## [v0.5]
 
 Main changes are: